Refactor the codebase
The codebase is now broken into 3 major language, python, java, and go.
Each language has its own .bzl file that lives in the language root dir.
For java it is `java/java.bzl`, for go it is `go/go.bzl`, and for python
it is `python/python.bzl`.

It is now discouraged for users to import from
//:bazel_integration_test.bzl. This has been made explicit to prevent
users from accidentally bringing in the dependencies of other languages.

Additionally I refactored some of the test cases to have less
indirection. It should be a little more clear what is happening in those
cases.

I also worked through some of the README.md files.